requirements.txt
streamlit==1.31.0
pandas==2.0.3
Pillow==10.0.0
requests==2.31.0
openpyxl==3.1.2
python-dotenv==1.0.1
numpy==1.24.3